Moves Resource unit tests to subdir to reflect changes in main namespace
