+++ tests/Unit/Resource/ContentTypeTest.php
@@ -7,7 +7,7 @@
  * @license   MIT
  */
 
-namespace Contentful\Tests\Unit;
+namespace Contentful\Tests\Unit\Resource;
 
 use Contentful\Management\Field\NumberField;
 use Contentful\Management\Field\SymbolField;
