+++ tests/Unit/Resource/EntryTest.php
@@ -7,7 +7,7 @@
  * @license   MIT
  */
 
-namespace Contentful\Tests\Unit;
+namespace Contentful\Tests\Unit\Resource;
 
 use Contentful\Link;
 use Contentful\Management\Resource\Entry;
