+++ tests/Unit/Resource/LocaleTest.php
@@ -7,7 +7,7 @@
  * @license   MIT
  */
 
-namespace Contentful\Tests\Unit;
+namespace Contentful\Tests\Unit\Resource;
 
 use Contentful\Management\Resource\Locale;
